price-tracker-master/price-tracker/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fastapi==0.110.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pip==24.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic_core==2.16.3
pydantic==2.6.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
thop

tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.29
urllib3==2.2.1
uvicorn==0.28.1
watchfiles==0.21.0
websockets==12.0
